+++ CHANGELOG.md
@@ -7,7 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Development V2
 
+### Changed
+- startup time tracking when autoplay is enabled starts from `SourceLoad` event
+- transition to `playing` state if collector is attached after player started playing 
+
 ## Development V1
+
+### Changed
+- startup time tracking when autoplay is enabled starts from `SourceLoaded` event
+- transition to `playing` state if collector is attached after player started playing 
 
 ## 1.24.1